* commit '059f699e4015b4d71cb94c480886aa5fcc665e49': (46 commits)
  doi htmllabel
  done
  bỏ htmllable
  doing
  .
  update nuget
  update version đẩy app 03/03/2021
  done
  done
  done alert
  done
  done alert
  .
  doing
  change giao diện tết về bình thường
  fix local DB
  update notice v2
  int => long
  .
  .
  ...

+++ src/BuildingBlocks/BA_MobileGPS.Entities/ResponeEntity/PaperItemHistoryModel.cs
@@ -14,7 +14,7 @@ namespace BA_MobileGPS.Entities.ResponeEntity
     {
         public Guid Id { get; set; }
         //public int FKCompanyID { get; set; }
-        public int FK_VehicleID { get; set; }
+        public long FK_VehicleID { get; set; }
         public Guid FK_PaperCategoryID { get; set; }
         public string PaperCategoryName { get; set; }
         //public string PaperNumber { get; set; }
